prevent Aboutdlg crash in case of wrong versiondata (DC++ BZR) add 1000 mbit upload speed (DC++ BZR) some fixes for resuming old .antifrag files (DC++ BZR) decreased OpenSSL libraries size fixed finished items' total size added option to place usercommands outside of submenu

+++ trunk/client/CryptoManager.h
@@ -42,7 +42,7 @@ public:
 	const string& getPk() const  { return pk; }
 	bool isExtended(const string& aLock) const { return strncmp(aLock.c_str(), "EXTENDEDPROTOCOL", 16) == 0; }
 
-	void decodeBZ2(const uint8_t* is, size_t sz, string& os) throw(CryptoException);
+	void decodeBZ2(const uint8_t* is, unsigned int sz, string& os) throw(CryptoException);
 
 	SSLSocket* getClientSocket(bool allowUntrusted) throw(SocketException);
 	SSLSocket* getServerSocket(bool allowUntrusted) throw(SocketException);
