+++ trunk/client/UploadManager.h
@@ -96,10 +96,10 @@ public:
 	 */
 	int64_t getRunningAverage();
 	
-	uint8_t getSlots() const { return (uint8_t)(max(SETTING(SLOTS), max(SETTING(HUB_SLOTS),0) * Client::getTotalCounts())); }
+	size_t getSlots() const { return max(SETTING(SLOTS), max(SETTING(HUB_SLOTS),0) * Client::getTotalCounts()); }
 
 	/** @return Number of free slots. */
-	uint8_t getFreeSlots() const { return (uint8_t)max((getSlots() - running), 0); }
+	size_t getFreeSlots() const { return max((getSlots() - running), (size_t)0); }
 	
 	/** @internal */
 	int getFreeExtraSlots() const { return max(SETTING(EXTRA_SLOTS) - getExtra(), 0); }
@@ -131,7 +131,7 @@ private:
 	bool isFireball;
 	bool isFileServer;
 	bool mThrottleEnable;
-	uint8_t running;
+	size_t running;
 
 	size_t mBytesSpokenFor, mUploadLimit, mCycleTime, mByteSlice;
 	uint64_t m_iHighSpeedStartTick;
@@ -146,7 +146,7 @@ private:
 	SlotMap connectingUsers;
 	UploadQueueItem::SlotQueue waitingUsers;
 
-	int addFailedUpload(const UserPtr& aUser, const string& file, int64_t pos, int64_t size);
+	size_t addFailedUpload(const UserPtr& aUser, const string& file, int64_t pos, int64_t size);
 	void notifyQueuedUsers();
 
 	void throttleSetup();
